began abstracting render systme

+++ src/nimgine/components.nim
@@ -2,8 +2,8 @@ import ecs
 
 type
     Position* = ref object of Component
-        x*, y*: int
+        x*, y*: float
     Dimensions* = ref object of Component
-        width*, height*: int
+        width*, height*: float
     Controllable* = ref object of Component
     RenderBlock* = ref object of Component
